Add class to execute submissions on test cases
The SubmissionExecutor runs tests on a Submission. Each time a test set
is run, an image is built, though this can be done higher in the call
chain when we figure that part out. Perhaps we can build the images upon
assignment creation, when a list of allowed languages is specified.

When each test case is run, a new container is created from the
appropriate image, and the test input file and submission file are
copied into the container. A command 'cmd', which should belong to the
Problem, is run. Right now, the default cmd stores the output of the
program in a file called 'output/program_output', and the
SubmissionExecutor reads the contents of this file to return as the
output of run_test.

Currently, the SubmissionExecutor runs a CPP program in var/submission.
The path is hardcoded, and is used as a default if no Submission is
provided to the run_tests method.